scripts/requirements-landing.txt
google-genai>=0.5.0
pillow>=10.3.0
